+++ visualization/app/codeCharta/codeCharta.model.ts
@@ -12,12 +12,6 @@ export interface NameDataPair {
 	content: ExportCCFile
 }
 
-export enum SearchPanelMode {
-	treeView = "treeView",
-	blacklist = "blacklist",
-	minimized = "minimized"
-}
-
 export enum LayoutAlgorithm {
 	SquarifiedTreeMap = "Squarified TreeMap",
 	StreetMap = "StreetMap",
@@ -146,9 +140,7 @@ export interface AppSettings {
 	isLoadingMap: boolean
 	isLoadingFile: boolean
 	sortingOrderAscending: boolean
-	searchPanelMode: SearchPanelMode
 	isAttributeSideBarVisible: boolean
-	panelSelection: PanelSelection
 	showMetricLabelNameValue: boolean
 	showMetricLabelNodeName: boolean
 	layoutAlgorithm: LayoutAlgorithm
@@ -399,12 +391,4 @@ export interface AppStatus {
 	hoveredBuildingPath: string | null
 	selectedBuildingId: number | null
 	rightClickedNodeData: RightClickedNodeData
-}
-
-export enum PanelSelection {
-	AREA_PANEL_OPEN = "AREA_PANEL_OPEN",
-	HEIGHT_PANEL_OPEN = "HEIGHT_PANEL_OPEN",
-	COLOR_PANEL_OPEN = "COLOR_PANEL_OPEN",
-	EDGE_PANEL_OPEN = "EDGE_PANEL_OPEN",
-	NONE = "NONE"
 }