+++ visualization/app/codeCharta/state/store/appSettings/panelSelection/panelSelection.splitter.ts
@@ -1,6 +1,0 @@
-import { setPanelSelection } from "./panelSelection.actions"
-import { PanelSelection } from "../../../../codeCharta.model"
-
-export function splitPanelSelectionAction(payload: PanelSelection) {
-	return setPanelSelection(payload)
-}