+++ visualization/app/codeCharta/ui/mapTreeView/mapTreeViewLevel/mapTreeViewLevel.component.ts
@@ -45,7 +45,6 @@ export class MapTreeViewLevelComponent implements OnInit {
 	}
 
 	openNodeContextMenu = $event => {
-		$event.stopPropagation()
 		$event.preventDefault()
 
 		this.store.dispatch(